[refactor(postResult, PostWinner)] void에서 String return으로 변경
