+++ src/main/java/racingcar/exception/dto/ErrorResponse.java
@@ -1,4 +1,4 @@
-package racingcar.exception;
+package racingcar.exception.dto;
 
 public class ErrorResponse {
