+++ src/main/java/racingcar/message/MessageResponse.java
@@ -5,4 +5,8 @@ public class MessageResponse {
 	public static void of(String message){
 		System.out.println(message);
 	}
+
+	public static void NonLineOf(String message){
+		System.out.print(message);
+	}
 }
